Add namespacing utils
